Slim down workspace-migration interceptors/utils
- Merge `getMetadataValidationUserFriendlyMessage` into
  `build-metadata-validation-error-payload.util.ts`. The picker only existed
  to feed `buildMetadataValidationErrorPayload`; both functions stay exported
  (the picker keeps its existing 6-test spec, now renamed alongside the
  merged source file).
- Move `translate-user-friendly-message-descriptors.util.ts` (and its spec)
  from `workspace-migration/interceptors/utils/` to `core-modules/i18n/utils/`.
  It's a generic descriptor-translation helper used by `useGraphQLErrorHandlerHook`
  for every `BaseGraphQLError`, not just the metadata-validation REST helper.
  Drops the back-reference to `FlatEntityValidationError` in favour of an
  inline `'userFriendlyMessage'` constant.

Net effect on `workspace-migration/interceptors/utils/`: 5 source files -> 3
(`build-metadata-validation-error-payload.util.ts`,
 `workspace-migration-builder-graphql-api-exception-handler.util.ts`,
 `workspace-migration-builder-rest-api-exception-handler.util.ts`).
